change Player to just Paddle, wip on reimagining how to render this paddle
