+++ public/js/managers/pixiManager.js
@@ -1,9 +1,10 @@
 import * as PIXI from 'pixi.js';
 
 import TextComponent from 'components/TextComponent';
-import PlayerComponent from 'components/PlayerComponent';
+import PaddleComponent from 'components/PaddleComponent';
 
 import { GAME_SIZE } from 'constants/sizes';
+import { PRIMARY_PLAYER_DEFAULT_POS, SECONDARY_PLAYER_DEFAULT_POS } from 'constants/positions';
 
 import { getCanvasContainer } from 'helpers/canvasHelper';
 
@@ -21,8 +22,12 @@ canvas.appendChild(app.view);
 
 /**
  * set up the components and elements that will show up on the screen
+ *
+ * todo: this will potentially grow to be unmaintainable, figure out a better solution
  */
 const setupApp = () => {
+  const stage = app.stage;
+
   // player count
   const playerCountText = new TextComponent('n connected player(s)', {
     position: {
@@ -30,12 +35,15 @@ const setupApp = () => {
       y: 20,
     },
   });
+  stage.addChild(playerCountText);
 
-  app.stage.addChild(playerCountText);
+  // opposing player
+  const secondaryPaddle = new PaddleComponent({position: SECONDARY_PLAYER_DEFAULT_POS});
+  stage.addChild(secondaryPaddle);
 
-  // ...
-  const primaryPlayer = new PlayerComponent();
-  app.stage.addChild(primaryPlayer.component);
+  // active player
+  const primaryPaddle = new PaddleComponent({position: PRIMARY_PLAYER_DEFAULT_POS});
+  stage.addChild(primaryPaddle);
 };
 
 // create singleton
